Modify files according CheckStyle(work in progress)
