+++ src/AI.md
@@ -1,6 +1,6 @@
-# TaskMaster and AI Integration
+# taskmaster and AI Integration
 
-TaskMaster incorporates AI tools to support and enhance the development process. These tools were utilized in specific ways to assist with productivity, troubleshooting, and improving the overall quality of the project. This document transparently declares the use of AI in developing this application.
+taskmaster incorporates AI tools to support and enhance the development process. These tools were utilized in specific ways to assist with productivity, troubleshooting, and improving the overall quality of the project. This document transparently declares the use of AI in developing this application.
 
 ## Use of AI
 
@@ -15,7 +15,7 @@ AI tools such as ChatGPT were used to:
 AI provided support in generating and reviewing alternative implementations for various methods, which helped improve the overall structure and readability of the code. For example:
 - Parsing logic was designed to support multiple date formats by leveraging AI-suggested solutions.
 - Error handling mechanisms were streamlined to ensure user-friendly messages and graceful handling of edge cases.
-
+- Helps in fixing issues with regard to the java coding standards and conventions.
 ### 3. Documentation Assistance
 AI tools assisted in drafting the initial versions of documentation files, including:
 - The `README.md`, which provides instructions for installation, usage, and examples of commands.
