--- src/main/java/TaskMaster/Commands/MarkCommand.java
@@ -1,32 +1,0 @@
-package TaskMaster.Commands;
-import TaskMaster.Storage.Storage;
-import TaskMaster.Utils.TaskList;
-import TaskMaster.Ui.Ui;
-import TaskMaster.Exceptions.TaskMasterException;
-import TaskMaster.Tasks.Task;
-/**
- * Command to mark a task as done.
- */
-public class MarkCommand extends Command {
-    private final int index;
-
-    /**
-     * Constructs a MarkCommand.
-     *
-     * @param index The index of the task to mark as done.
-     */
-    public MarkCommand(int index) {
-        this.index = index;
-    }
-
-    @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) throws TaskMasterException {
-        if (index <= 0 || index > tasks.getTasks().size()) {
-            throw new TaskMasterException("Task index out of range.");
-        }
-        Task task = tasks.getTasks().get(index - 1);
-        task.markAsDone();
-        ui.show("Nice! I've marked this task as done:");
-        ui.show("  " + task);
-    }
-}
