+++ src/main/java/taskmaster/storage/Storage.java
@@ -1,0 +1,75 @@
+package taskmaster.storage;
+
+import java.io.File;
+import java.io.FileWriter;
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.Scanner;
+
+import taskmaster.exceptions.TaskMasterException;
+import taskmaster.parser.Parser;
+import taskmaster.tasks.Task;
+
+/**
+ * Handles loading tasks from and saving tasks to a file.
+ */
+public class Storage {
+    private final String filePath;
+
+    /**
+     * Constructs a new storage instance with the specified file path.
+     *
+     * @param filePath The path to the file for storing tasks.
+     */
+    public Storage(String filePath) {
+        this.filePath = filePath;
+    }
+
+    /**
+     * Loads tasks from the file.
+     *
+     * @return A list of tasks loaded from the file.
+     * @throws IOException If there is an error reading the file.
+     */
+    public ArrayList<Task> load() throws IOException {
+        ArrayList<Task> tasks = new ArrayList<>();
+        File file = new File(filePath);
+
+        if (!file.exists()) {
+            if (file.getParentFile() != null) {
+                file.getParentFile().mkdirs();
+            }
+            boolean created = file.createNewFile();
+            if (created) {
+                System.out.println("Created new task storage file: " + filePath);
+            }
+        }
+
+        try (Scanner scanner = new Scanner(file)) {
+            while (scanner.hasNextLine()) {
+                String line = scanner.nextLine();
+                try {
+                    tasks.add(Parser.parseTask(line));
+                } catch (TaskMasterException e) {
+                    System.out.println("Skipping invalid task in file: " + line);
+                }
+            }
+        }
+
+        return tasks;
+    }
+
+    /**
+     * Saves the given list of tasks to the file.
+     *
+     * @param tasks The list of tasks to save.
+     * @throws IOException If there is an error writing to the file.
+     */
+    public void save(ArrayList<Task> tasks) throws IOException {
+        try (FileWriter writer = new FileWriter(filePath)) {
+            for (Task task : tasks) {
+                writer.write(task.save() + "\n");
+            }
+        }
+    }
+}
